add context validation at alterarUsuario

+++ projeto-final/backend/resolvers/Mutation/usuario.js
@@ -72,7 +72,7 @@ const mutations = {
             const usuario = await obterUsuario(_, { filtro });
             if (usuario) {
                 const { id } = usuario
-                if (dados.perfis) {
+                if (context.admin && dados.perfis) {
                     await db('usuarios_perfis')
                         .where({ usuario_id: id }).delete()
